agent_service/requirements.txt
langchain
langchain-community
mistralai
beautifulsoup4
requests
pydantic
python-dotenv